pyproject.toml
[tool.poetry]
name = "schoolsyst_api"
version = "0.1.0"
description = "schoolsyst.com's web API, built with FastAPI."
authors = ["Ewen Le Bihan <hey@ewen.works>"]
license = "AGPL-3.0-only"
repository = "https://github.com/schoolsyst/api-next"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.60.1"
ujson = "^3.0.0"
email_validator = "^1.1.1"
requests = "^2.24.0"
aiofiles = "^0.5.0"
jinja2 = "^2.11.2"
itsdangerous = "^1.1.0"
graphene = "^2.1.8"
uvicorn = "^0.11.6"
python-multipart = "^0.0.5"
passlib = {extras = ["argon2"], version = "^1.7.2"}
python-dotenv = "^0.14.0"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
toml = "^0.10.1"
orjson = "^3.2.2"
parse = "^1.16.0"
python-arango = "^5.4.0"
zxcvbn = "^4.4.28"
typed_dotenv = "^1.0.1"
isodate = "^0.6.0"
nanoid = "^2.0.0"
python-slugify = "^4.0.1"
fastapi_utils = "^0.2.1"
fastapi-etag = "^0.2.2"
chevron = "^0.13.1"
yagmail = "^0.11.224"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pydeps = "^1.9.3"
pylint = "^2.5.3"
rope = "^0.17.0"
pre-commit = "^2.6.0"
pytest = "^6.0.1"
pytest-cov = "^2.10.0"
autoflake = "^1.3.1"
flake8 = "^3.8.3"
codecov = "^2.1.8"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
